Enforce PRP's 2.2 & 2.3 / Expose state & result
Immediately fulfill/reject promise, in [[Resolve]](promise, x), if x is already
fulfilled/rejected. This is according to steps 2.2 & 2.3 as stated in the
Promise Resolution Procedure. (Note that skipping these and just using step 2.1
as a 'generalization' of 2, goes unnoticed by the test suite). Implement getState()
and getResult() on promise.